+++ ares-exec-plugin/src/main/kotlin/io/github/sergkhram/tasks/ExecutionWReportTask.kt
@@ -3,7 +3,7 @@ package io.github.sergkhram.tasks
 import io.github.sergkhram.configuration.Configuration
 import io.github.sergkhram.configuration.ConfigurationExtension
 import io.github.sergkhram.configuration.ExecuteBy
-import io.github.sergkhram.helpers.CustomException
+import io.github.sergkhram.getPropertyExecuteBy
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.VerificationTask
@@ -18,12 +18,7 @@ open class ExecutionWReportTask: DefaultTask(), VerificationTask {
         if(!Configuration.executionIgnoreFailures) {
             val conf = project.extensions.getByName("ares") as? ConfigurationExtension
                 ?: ConfigurationExtension(project)
-            val executeBy = try {
-                System.getProperty("executeBy")?.let { ExecuteBy.valueOf(it) }
-                    ?: conf.testExecutionBlock?.executeBy?.let { ExecuteBy.valueOf(it) }
-            } catch (e: IllegalArgumentException) {
-                throw CustomException("There is no chosen executeBy variant, only these values are supported : ${ExecuteBy.values().map {it.name}}")
-            }
+            val executeBy = getPropertyExecuteBy(conf)
             executeBy?.let {
                 when(it) {
                     ExecuteBy.MARATHON -> {
