+++ ares-exec-plugin/src/main/kotlin/io/github/sergkhram/tasks/taskActions.kt
@@ -1,12 +1,14 @@
 package io.github.sergkhram.tasks
 
 import com.malinskiy.marathon.MarathonRunTask
+import com.malinskiy.marathon.android.configuration.AllureConfiguration
+import com.malinskiy.marathon.device.DeviceFeature
+import io.github.sergkhram.*
 import io.github.sergkhram.configuration.Configuration
+import io.github.sergkhram.configuration.ConfigurationExtension
+import io.github.sergkhram.configuration.MarathonBlock
 import io.github.sergkhram.configuration.ScreenRecordAttachment
 import io.github.sergkhram.helpers.CustomException
-import io.github.sergkhram.logger
-import io.github.sergkhram.marathonGifDir
-import io.github.sergkhram.marathonVideoDir
 import org.gradle.api.GradleException
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.get
@@ -17,11 +19,7 @@ fun marathonAfterExecutionAction(project: Project) {
     val marathonTaskName = "marathon${buildType.capitalize()}AndroidTest"
     val marathonTask = project.tasks.get(marathonTaskName) as MarathonRunTask
     logger.info("Getting retry count config value")
-    val retryCount: Int = marathonTask.marathonExtension.get().retryStrategy?.let { retryStrategy ->
-        retryStrategy.fixedQuota?.let { fixedQuota ->
-            fixedQuota.retryPerTestQuota
-        }
-    } ?: 3
+    val retryCount: Int = marathonTask.getRetryCountValue()
     logger.info("Retry count value is $retryCount")
     val reportDirectory = Configuration.getReportDirectory(project.projectDir.path)
     logger.info("ReportDirectory is $reportDirectory")
@@ -30,7 +28,7 @@ fun marathonAfterExecutionAction(project: Project) {
     val marathonVideoDir = marathonVideoDir(reportDirectory)
     val marathonGifDir = marathonGifDir(reportDirectory)
     if(marathonVideoDir.exists() || marathonGifDir.exists()) {
-        if (!currentMarathonVideosDir.exists()) throw CustomException("There is no $currentMarathonVideosDir directory. Check the attachmentType property")
+        if (!currentMarathonVideosDir.exists()) throw CustomException("There is no $currentMarathonVideosDir directory. Check the screenRecordType property")
         val failedTests = mutableMapOf<String, Int>()
         currentMarathonVideosDir.listFiles()!!.forEach { shardingStrategy ->
             shardingStrategy.listFiles()!!.forEach { deviceDir ->
@@ -57,4 +55,34 @@ fun failedTestsExceptionAction(executor: String) {
     val message = "At least one test failed during all attempts by $executor. See build/allure-results directory or generate report using allure gradle plugin"
     logger.info(message)
     throw GradleException(message)
+}
+
+fun MarathonRunTask.enableAllureForMarathon() {
+    this.marathonExtension.get().apply {
+        if(this.allureConfiguration==null || !this.allureConfiguration!!.enabled) {
+            val allureResultsDir: String? = this.allureConfiguration?.relativeResultsDirectory
+            this.allureConfiguration = AllureConfiguration().apply {
+                this.enabled = true
+                allureResultsDir?.let { resDir -> this.relativeResultsDirectory = resDir }
+            }
+        }
+    }
+}
+
+fun MarathonRunTask.getRetryCountValue(): Int {
+    return this.marathonExtension.get().retryStrategy?.let { retryStrategy ->
+        retryStrategy.fixedQuota?.let { fixedQuota ->
+            fixedQuota.retryPerTestQuota
+        }
+    } ?: 3
+}
+
+fun ConfigurationExtension.setScreenRecordTypeFromMarathon(marathonTask: MarathonRunTask) {
+    val marathonVideoConfiguration = marathonTask.marathonExtension.get().screenRecordConfiguration?.preferableRecorderType?.name
+        ?: DeviceFeature.VIDEO.name
+    if(this.marathonBlock == null) {
+        this.marathonBlock = MarathonBlock(screenRecordType = marathonVideoConfiguration)
+    } else {
+        this.marathonBlock!!.screenRecordType = marathonVideoConfiguration
+    }
 }
