+++ ares-plugin/src/main/kotlin/io/github/sergkhram/configuration/AllureBlock.kt
