+++ ares-plugin/src/main/kotlin/io/github/sergkhram/configuration/ConfigurationExtension.kt
