+++ ares-plugin/src/main/kotlin/io/github/sergkhram/helpers/conditions.kt
